requirements.txt
appdirs==1.4.4
build==1.2.2
buildozer==1.5.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Cython==3.0.11
distlib==0.3.8
docutils==0.21.2
filelock==3.16.0
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
Kivy==2.3.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.3
proto-plus==1.24.0
protobuf==5.28.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyjnius==1.6.1
pyparsing==3.1.4
pyproject_hooks==1.1.0
pytesseract==0.3.13
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
setuptools==75.1.0
sh==1.14.3
toml==0.10.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.4
